app/requirements.txt
asgiref==3.3.4
autopep8==1.5.7
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
dj-database-url==0.5.0
Django==3.2.4
django-cors-headers==3.7.0
django-debug-toolbar==3.2.1
django-dotenv==1.4.2
django-environ==0.4.5
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
google-api-core==1.22.2
google-api-python-client==1.12.1
google-auth==1.31.0
google-auth-httplib2==0.0.4
googleapis-common-protos==1.52.0
gunicorn==20.1.0
httplib2==0.19.1
idna==2.10
protobuf==3.17.3
psycopg2-binary==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
PyJWT==2.1.0
pyparsing==2.4.7
python-decouple==3.4
python-dotenv==0.17.1
pytz==2021.1
requests==2.25.1
rsa==4.7.2
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.5
whitenoise==5.2.0
